Added a __main__.py entrypoint
With a `config.yaml` there is no need to import this into a
`chatbot.py`. Instead just configure it in the yaml and then call
`Legobot`, ie `python3 Legobot`. I'd like to add some basic argv
support, but that's sorta out of scope for this I think, also would hate
to do that effort only for it to be rejected because the use of a
`__main__.py` is unwanted.